+++ formiz-core/src/test/java/org/formiz/core/expr/test1/EnglishExpressionTest.java
@@ -39,10 +39,10 @@ import java.util.Map;
 import org.formiz.core.expr.IContext;
 import org.formiz.core.expr.IExpression;
 import org.formiz.core.expr.impl.ContextBuilder;
-import org.formiz.core.expr.impl.ElContext;
-import org.formiz.core.expr.impl.ElExpressionParser;
 import org.formiz.core.expr.impl.EnglishExpressionParser;
 import org.formiz.core.expr.impl.ParseException;
+import org.formiz.core.expr.spel.ElContext;
+import org.formiz.core.expr.spel.ElExpressionParser;
 import org.formiz.core.expr.test1.model.Address;
 import org.formiz.core.expr.test1.model.Functions;
 import org.junit.Assert;
